backend/requirements.txt
# Web framework and server
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7

# Environment and configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Semantic similarity
sentence-transformers==2.2.2
numpy==1.25.2
scikit-learn==1.2.2

# Text processing and security
bleach==6.0.0
spacy==3.6.1

# Logging
loguru==0.7.0
